event: removed on-event
The events mechanism was changed so that event dispatch is
no longer tied to the listener class hierarchy.  This association
created problems with inheritance, e.g.:

 * the instance is shared both by superclass and subclass code,
   so if one adds it as a listener to some speaker, the other
   will intercept speaker's events while, from its point of view,
   no listener was added to it;
 * remove-listener could remove an instance added twice, by both
   superclass and subclass code (reference counting would not work
   here because of the former example);
 * a subclass's on-event would inhibit a superclass's one if defined
   (if this was an isolated problem, we could solve it by calling
    the next method, or using :after, or using progn method combination)

The problem is using instances as listeners, while we want a finer
granularity distinguishing superclass use from subclass use.

A solution is to get rid of on-event.  Listeners then are functions
taking a single parameter, the event.  It is possible, of course,
to delegate to a generic function taking more parameters.  The
association between event dispatch and class hierarchy is then
removed from the framework, while the particular superclass-subclass
protocols may introduce it themselves explicitly.